+++ R/Lib_PROSAIL_HybridInversion.R
@@ -954,7 +954,7 @@ train_prosail_inversion <- function(InputPROSAIL = NULL, BRF_LUT = NULL,
     ### 2- PRODUCE BRF from InputPROSAIL & ddefault spectral sampling = 1nm  ###
     ### == == == == == == == == == == == == == == == == == == == == == == == ###
     # define default SpecPROSPECT, SpecSOIL and SpecATM if undefined
-    if (is.null(SpecPROSPECT)) SpecPROSPECT <- prospect::SpecPROSPECT_FullRange
+    if (is.null(SpecPROSPECT)) SpecPROSPECT <- prospect::spec_prospect_full_range
     if (is.null(SpecSOIL)) SpecSOIL <- prosail::SpecSOIL
     if (is.null(SpecATM)) SpecATM <- prosail::SpecATM
     # check if same spectral sampling for all key variables